+++ OpenGL-basico/Utils.h
@@ -2,4 +2,5 @@
 #include <GL/gl.h>
 #include "Colors.h" 
 
-void drawCube(int x, int y, int z, float size, Colors color);
+//void drawCube(int x, int y, int z, float size, Colors color);
+void drawCube(float x, float y, float z, float size, Colors color);
